P17: created interactable button that opens specific door
