+++ Assets/Prefabs/Interactables.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 57b0b4eeb84e3f54cb19a931e9d3b320
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
