+++ Assets/Prefabs/Interactables/I03_Button.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7eb300012f7b41047ac5054d3b81bf9e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
