+++ Assets/Scripts/Interactables/I01_Key.cs
@@ -2,27 +2,27 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Door : MonoBehaviour, IInteractable
+public class I01_Key : MonoBehaviour, IInteractable
 {
     [SerializeField] private string m_id;
-    [SerializeField] private string m_prompt;
+    [SerializeField] private string m_prompt = "Grab Key";
 
     public string ID => m_id;
     public string InteractionPromt => m_prompt;
 
     public bool Interact(PlayerInteract interactor)
     {
-        Debug.Log("Door Opened");
+        Debug.Log("Key Grabbed");
         return true;
     }
 
     public void OnFocus()
     {
-        Debug.Log("Focus on " + ID);
+        return;
     }
 
     public void OnLoseFocus()
     {
-        Debug.Log("Lose Focus on " + ID);
+        return;
     }
 }
