+++ Assets/Scripts/Interactables/I01_Key.cs.meta
