+++ Assets/Scripts/Interactables/I03_Button.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3abf8eab1de3b75428951ba2564a19f6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
